+++ app/Http/Controllers/RaidController.php
@@ -3,6 +3,7 @@
 namespace heroisNW\Http\Controllers;
 
 use Illuminate\Http\Request;
+use heroisNW\Http\Requests\RaidRequest;
 use heroisNW\Raid;
 use heroisNW\Personagem;
 
@@ -43,7 +44,7 @@ class RaidController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(RaidRequest $request)
     {
         $raid = Raid::create($request->all());
         $raid->personagens()->sync($request->input('personagem_array'));
@@ -59,7 +60,15 @@ class RaidController extends Controller
      */
     public function show($id)
     {
-        //
+        $raid = Raid::find($id);
+
+        if (is_null($raid)) {
+            return redirect()->action('RaidController@index')->withErrors('Raid não encontrada.');
+        } else {
+            return view('raid.raidShow')
+                ->with('titulo', 'Raids - Detalhes')
+                ->with('raid', $raid);
+        }
     }
 
     /**
@@ -70,7 +79,18 @@ class RaidController extends Controller
      */
     public function edit($id)
     {
-        //
+        $raid = Raid::find($id);
+
+        if (is_null($raid)) {
+            return redirect()->action('RaidController@index')->withErrors('Raid não encontrada.');
+        } else {
+            $personagens = Personagem::all();
+
+            return view('raid.raidForm')
+                ->with('titulo', 'Raids - Alteração')
+                ->with('raid', $raid)
+                ->with('personagens', $personagens);
+        }
     }
 
     /**
@@ -80,9 +100,14 @@ class RaidController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(RaidRequest $request, $id)
     {
-        //
+        $raid = Raid::find($id);
+        $raid->fill($request->all());
+        $raid->personagens()->sync($request->input('personagem_array'));
+        $raid->save();
+
+        return redirect()->action('RaidController@index');
     }
 
     /**
@@ -94,9 +119,14 @@ class RaidController extends Controller
     public function destroy($id)
     {
         $raid = Raid::find($id);
-        $raid->personagens()->sync([]);
-        $raid->delete();
 
-        return redirect()->action('RaidController@index');
+        if (is_null($raid)) {
+            return redirect()->action('RaidController@index')->withErrors('Raid não encontrada.');
+        } else {
+            $raid->personagens()->sync([]);
+            $raid->delete();
+
+            return redirect()->action('RaidController@index');
+        }
     }
 }
